Refactor batching logic for collections, tracks, and users
