+++ packages/common/src/api/tan-query/useUsers.ts
@@ -1,54 +1,16 @@
-import { Id, OptionalId } from '@audius/sdk'
-import { QueryClient, useQueries, useQueryClient } from '@tanstack/react-query'
-import { create, keyResolver, windowScheduler } from '@yornaath/batshit'
+import { useQueries, useQueryClient } from '@tanstack/react-query'
 import { useDispatch } from 'react-redux'
-import { Dispatch } from 'redux'
 
-import { userMetadataListFromSDK } from '~/adapters/user'
 import { useAudiusQueryContext } from '~/audius-query'
 import { ID } from '~/models/Identifiers'
 import { UserMetadata } from '~/models/User'
-import { removeNullable } from '~/utils/typeUtils'
 
+import { getUsersBatcher } from './batchers/getUsersBatcher'
 import { QUERY_KEYS } from './queryKeys'
 import { QueryOptions } from './types'
 import { useCurrentUserId } from './useCurrentUserId'
 import { getUserQueryKey } from './useUser'
 import { combineQueryResults } from './utils/combineQueryResults'
-import { primeUserData } from './utils/primeUserData'
-
-type BatchContext = {
-  sdk: any
-  currentUserId: ID | null | undefined
-  queryClient: QueryClient
-  dispatch: Dispatch
-}
-
-type BatchQuery = {
-  id: ID
-  context: BatchContext
-}
-
-const getUsersBatcher = create({
-  fetcher: async (queries: BatchQuery[]): Promise<UserMetadata[]> => {
-    // Hack because batshit doesn't support context properly
-    const { sdk, currentUserId, queryClient, dispatch } = queries[0].context
-    if (!queries.length) return []
-    const ids = queries.map((q) => q.id)
-    const { data } = await sdk.full.users.getBulkUsers({
-      id: ids.map((id: ID) => Id.parse(id)).filter(removeNullable),
-      userId: OptionalId.parse(currentUserId)
-    })
-
-    // prime the users in the cache
-    const users = userMetadataListFromSDK(data)
-    primeUserData({ users, queryClient, dispatch })
-
-    return users
-  },
-  resolver: keyResolver('user_id'),
-  scheduler: windowScheduler(10)
-})
 
 export const getUsersQueryKey = (userIds: ID[] | null | undefined) => [
   QUERY_KEYS.users,
